requirements.txt
fastapi==0.115.10
uvicorn==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
pillow==11.1.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.6